requirements.txt
anyio==3.7.1
asgiref==3.7.2
bidict==0.22.1
certifi==2023.5.7
cffi==1.15.1
cryptography==41.0.2
Django==4.2.3
dnspython==2.4.0
eventlet==0.33.3
exceptiongroup==1.1.2
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
packaging==23.1
psutil==5.9.5
pycparser==2.21
python-engineio==4.5.1
python-socketio==5.8.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
